+++ java/src/ru/koshakmine/icstd/ui/elements/ImageElement.java
@@ -5,7 +5,7 @@ import org.mozilla.javascript.ScriptableObject;
 public class ImageElement extends WindowElementSize {
     protected String bitmap;
 
-    public ImageElement(String bitmap, int x, int y, int width, int height) {
+    public ImageElement(String bitmap, float x, float y, int width, int height) {
         super("image", x, y, width, height);
         this.bitmap = bitmap;
     }
